separate out github authentication methods
